+++ src/main/java/io/dpopkov/knowthenix/rest/model/response/ErrorMessages.java
@@ -1,0 +1,15 @@
+package io.dpopkov.knowthenix.rest.model.response;
+
+public enum ErrorMessages {
+    MISSING_REQUIRED_FIELD("Missing required field. Please check documentation for required fields");
+
+    private final String errorMessage;
+
+    ErrorMessages(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+}