+++ src/main/java/io/dpopkov/knowthenix/services/dto/BaseDto.java
@@ -2,9 +2,11 @@ package io.dpopkov.knowthenix.services.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
+@ToString
 public class BaseDto {
     private Long id;
 }