+++ src/main/java/io/dpopkov/knowthenix/services/dto/CategoryDto.java
@@ -1,14 +1,12 @@
 package io.dpopkov.knowthenix.services.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@ToString(callSuper = true)
 public class CategoryDto extends BaseDto {
     private String name;
     private String description;