+++ history.md
@@ -1,6 +1,7 @@
-*# History of implementation
+# History of implementation
 
 (in reversed order)
 
+* CRUD for Category: controller.
 * CRUD for Category: service.
 * CRUD for Category: entity and repository.